# api/api.yaml
openapi: 3.0.0
info:
  title: Simple Items API
  version: 1.0.0

paths:
  /session/new:
    post:
      summary: Initiate new session
      parameters:
        - name: X-Idempotency-Token
          in: header
          schema:
            type: string
          required: true
      responses:
        '200':
          description: New session response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewSessionResponse'

  /session/submit:
    post:
      summary: Submit OAuth2 code
      parameters:
        - name: X-Idempotency-Token
          in: header
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                session_id:
                  type: string
              required:
                - code
                - session_id
      responses:
        '200':
          description: Validated session response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitSessionResponse'
        '400':
          description: Bad request

  /session/info:
    get:
      summary: Get session info
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Session info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionInfo'
        '400':
          description: Bad request
        '401':
          description: Unauthorized

components:
  schemas:
    NewSessionResponse:
      type: object
      properties:
        session_id:
          type: string
        authorization_endpoint:
          type: string
        response_type:
          type: string
        client_id:
          type: string
        scope:
          type: array
          items:
            type: string
        redirect_uri:
          type: string
        state:
          type: string
        nonce:
          type: string
      required:
        - session_id
        - authorization_endpoint
        - response_type
        - client_id
        - scope
        - redirect_uri
        - state
        - nonce

    SubmitSessionResponse:
      type: object
      properties:
        token:
          type: string
      required:
        - token

    SessionInfo:
      type: object